+++ .ci/tests.sh
@@ -3,8 +3,8 @@ set -e
 
 echo "👉  Building Test Image"
 TEST_TAG=$(echo "$TAG" | sed -e 's,/,\\\/,g')
-sed -i -e "s/{IMAGE_TAG}/$TEST_TAG/g" TestSuite/Dockerfile
-docker build TestSuite -t tests
+sed -i -e "s/{IMAGE_TAG}/$TEST_TAG/g" Tests/Dockerfile
+docker build Tests -t tests
 
 echo "👉  Running Test Suite"
 docker run tests