+++ Tests/ImageTools/Sources/ImageTools/ImageTools.swift
@@ -2,13 +2,13 @@ import CCurl
 import Foundation
 
 ///
-/// The tools to check the integrity of the image.
+/// A class to get information about a docker-swift-apns image.
 ///
 
-struct ImageTools {
+class ImageTools {
 
-    /// Determines the libcurl version string.
-    var cURLVersion: String {
+    /// Determines the libcurl version.
+    var cURLVersion: Version {
 
         let versionNum = curl_version_info(CURLVERSION_FOURTH)!.pointee.version_num
 
@@ -16,11 +16,11 @@ struct ImageTools {
         let minor = (versionNum & 0b000000001111111100000000) >> 8
         let patch = (versionNum & 0b000000000000000011111111)
 
-        return [major, minor, patch].map(String.init).joined(separator: ".")
+        return Version(major: major, minor: minor, patch: patch)
 
     }
 
-    /// Determines whether libcurl was built with HTTP/2 support.
+    /// Whether libcurl was built with HTTP/2 support.
     var hasHTTP2Support: Bool {
 
         let version = curl_version_info(CURLVERSION_FOURTH)!.pointee